requirements.txt
argcomplete==1.12.2
asgiref==3.3.1
asonic==2.0.0
cachetools==4.2.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.4
google-api-core==1.24.1
google-api-python-client==1.12.8
google-auth==1.24.0
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
googleapis-common-protos==1.52.0
httplib2==0.18.1
idna==2.10
image==1.5.33
importlib-metadata==3.3.0
oauthlib==3.1.0
pdf2image==1.14.0
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytesseract==0.3.7
pytz==2020.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
zipp==3.4.0
